Add HTML and CSS Files
Began work on the Landing Page project. Thus far I've added the
index as well as the style sheet and done some work here or there
as a form of self-checking before I get too deep. This time I'm
feeling VERY confident and actually excited to progress on this
project.